WIP: Separate endless scrolling and paging into their own fragments - Re-implement features
