+++ app/src/main/java/me/devsaki/hentoid/abstracts/DownloadsFragment.java
@@ -3,6 +3,7 @@ package me.devsaki.hentoid.abstracts;
 import android.animation.ObjectAnimator;
 import android.animation.ValueAnimator;
 import android.app.NotificationManager;
+import android.app.SearchManager;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
@@ -11,8 +12,10 @@ import android.os.Bundle;
 import android.os.Handler;
 import android.os.Parcelable;
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.v4.view.GravityCompat;
+import android.support.v4.view.MenuItemCompat;
 import android.support.v4.widget.DrawerLayout;
-import android.support.v4.widget.DrawerLayout.DrawerListener;
 import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -30,6 +33,9 @@ import android.widget.ImageButton;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
+import org.greenrobot.eventbus.Subscribe;
+import org.greenrobot.eventbus.ThreadMode;
+
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
@@ -38,6 +44,7 @@ import me.devsaki.hentoid.HentoidApp;
 import me.devsaki.hentoid.R;
 import me.devsaki.hentoid.activities.ImportActivity;
 import me.devsaki.hentoid.adapters.ContentAdapter;
+import me.devsaki.hentoid.adapters.ContentAdapter.ContentsWipedListener;
 import me.devsaki.hentoid.database.SearchContent;
 import me.devsaki.hentoid.database.SearchContent.ContentListener;
 import me.devsaki.hentoid.database.domains.Content;
@@ -57,45 +64,46 @@ import static me.devsaki.hentoid.util.Helper.DURATION.LONG;
  * Common elements for use by EndlessFragment and PagerFragment
  */
 public abstract class DownloadsFragment extends BaseFragment implements ContentListener,
-        DrawerListener, ItemSelectListener {
-    protected static final int SHOW_LOADING = 1;
-    protected static final int SHOW_BLANK = 2;
+        ContentsWipedListener, ItemSelectListener {
     protected static final int SHOW_RESULT = 3;
-    protected static final String LIST_STATE_KEY = "list_state";
+    private static final int SHOW_LOADING = 1;
+    private static final int SHOW_BLANK = 2;
+    private static final String LIST_STATE_KEY = "list_state";
 
     private static final String TAG = LogHelper.makeLogTag(DownloadsFragment.class);
 
     protected static String query = "";
-    protected final Handler searchHandler = new Handler();
+    private final Handler searchHandler = new Handler();
     protected Context mContext;
-    protected SharedPreferences prefs;
-    protected int order;
-    protected boolean orderUpdated;
     protected int qtyPages;
     protected int currentPage = 1;
     protected ContentAdapter mAdapter;
-    protected DrawerLayout mDrawerLayout;
-    protected int mDrawerState;
     protected LinearLayoutManager llm;
-    protected Parcelable mListState;
     protected RecyclerView mListView;
-    protected Button btnPage;
     protected List<Content> contents;
     protected List<Content> result = new ArrayList<>();
     protected SearchContent search;
-    protected SearchView searchView;
     protected LinearLayout toolTip;
-    protected SwipeRefreshLayout refreshLayout;
     protected Toolbar toolbar;
     protected boolean newContent;
     protected boolean override;
     protected boolean isLastPage;
     protected boolean isLoaded;
-    protected boolean isSelected;
-    protected boolean selectTrigger = false;
-    protected ActionMode mActionMode;
-    // Called when the action mode is created; startActionMode() was called
-    protected final ActionMode.Callback mActionModeCallback = new ActionMode.Callback() {
+    private ActionMode mActionMode;
+    private int mDrawerState;
+    private DrawerLayout mDrawerLayout;
+    private boolean shouldHide;
+    private MenuItem searchMenu;
+    private Parcelable mListState;
+    private Button btnPage;
+    private SearchView searchView;
+    private SwipeRefreshLayout refreshLayout;
+    private boolean orderUpdated;
+    private boolean isSelected;
+    private boolean selectTrigger = false;
+    // Called when the action mode is created; startActionMode() was called.
+    private final ActionMode.Callback mActionModeCallback = new ActionMode.Callback() {
+        // Called when action mode is first created.
         @Override
         public boolean onCreateActionMode(ActionMode mode, Menu menu) {
             // Inflate a menu resource providing context menu items
@@ -105,8 +113,8 @@ public abstract class DownloadsFragment extends BaseFragment implements ContentL
             return true;
         }
 
-        // Called each time the action mode is shown. Always called after onCreateActionMode, but
-        // may be called multiple times if the mode is invalidated.
+        // Called each time the action mode is shown. Always called after onCreateActionMode,
+        // but may be called multiple times if the mode is invalidated.
         @Override
         public boolean onPrepareActionMode(ActionMode mode, Menu menu) {
             menu.findItem(R.id.action_delete).setVisible(!selectTrigger);
@@ -116,7 +124,7 @@ public abstract class DownloadsFragment extends BaseFragment implements ContentL
             return true;
         }
 
-        // Called when the user selects a contextual menu item
+        // Called when the user selects a contextual menu item.
         @Override
         public boolean onActionItemClicked(ActionMode mode, MenuItem item) {
             switch (item.getItemId()) {
@@ -140,13 +148,16 @@ public abstract class DownloadsFragment extends BaseFragment implements ContentL
             }
         }
 
-        // Called when the user exits the action mode
+        // Called when the user exits the action mode.
         @Override
         public void onDestroyActionMode(ActionMode mode) {
             clearSelection();
             mActionMode = null;
         }
     };
+    private SharedPreferences prefs;
+    private int order;
+    private long backButtonPressed;
     private String settingDir;
     private TextView loadingText;
     private TextView emptyText;
@@ -154,169 +165,18 @@ public abstract class DownloadsFragment extends BaseFragment implements ContentL
     private ObjectAnimator animator;
 
     @Override
-    public boolean onBackPressed() {
-        return false;
-    }
+    public void onResume() {
+        super.onResume();
 
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
+        checkPermissions();
 
-        setRetainInstance(true);
-        setHasOptionsMenu(true);
-
-        mContext = getContext();
-
-        prefs = HentoidApp.getSharedPrefs();
-
-        settingDir = prefs.getString(Consts.SETTINGS_FOLDER, "");
-
-        order = prefs.getInt(
-                ConstsPrefs.PREF_ORDER_CONTENT_LISTS, ConstsPrefs.PREF_ORDER_CONTENT_ALPHABETIC);
-
-        qtyPages = Integer.parseInt(
-                prefs.getString(
-                        ConstsPrefs.PREF_QUANTITY_PER_PAGE_LISTS,
-                        ConstsPrefs.PREF_QUANTITY_PER_PAGE_DEFAULT + ""));
-    }
-
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, final ViewGroup container,
-                             Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.fragment_downloads, container, false);
-
-        initUI(rootView);
-        attachScrollListener();
-        attachOnClickListeners(rootView);
-
-        return rootView;
-    }
-
-    protected abstract void attachScrollListener();
-
-    private void attachOnClickListeners(View rootView) {
-        attachPrevious(rootView);
-        attachNext(rootView);
-        attachRefresh(rootView);
-
-        toolTip.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                commitRefresh();
-            }
-        });
-
-        refreshLayout.setEnabled(false);
-        refreshLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
-            @Override
-            public void onRefresh() {
-                commitRefresh();
-            }
-        });
-    }
-
-    private void attachPrevious(View rootView) {
-        ImageButton btnPrevious = (ImageButton) rootView.findViewById(R.id.btnPrevious);
-        btnPrevious.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (currentPage > 1 && isLoaded) {
-                    currentPage--;
-                    update();
-                } else if (qtyPages > 0 && isLoaded) {
-                    Helper.toast(mContext, R.string.not_previous_page);
-                } else {
-                    LogHelper.d(TAG, R.string.not_limit_per_page);
-                }
-            }
-        });
-    }
-
-    private void attachNext(View rootView) {
-        ImageButton btnNext = (ImageButton) rootView.findViewById(R.id.btnNext);
-        btnNext.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (qtyPages <= 0) {
-                    LogHelper.d(TAG, R.string.not_limit_per_page);
-                } else {
-                    if (!isLastPage && isLoaded) {
-                        currentPage++;
-                        update();
-                    } else if (isLastPage) {
-                        Helper.toast(mContext, R.string.not_next_page);
-                    }
-                }
-            }
-        });
-    }
-
-    protected abstract void attachRefresh(View rootView);
-
-    protected void clearQuery(int option) {
-        LogHelper.d(TAG, "Clearing query with option: " + option);
-        if (option == 1) {
-            if (searchView != null) {
-                searchView.clearFocus();
-                searchView.setIconified(true);
-            }
+        if (mListState != null) {
+            llm.onRestoreInstanceState(mListState);
         }
-        setQuery(query = "");
-        update();
-    }
-
-    private void commitRefresh() {
-        toolTip.setVisibility(View.GONE);
-        refreshLayout.setRefreshing(false);
-        refreshLayout.setEnabled(false);
-        newContent = false;
-        mAdapter.updateContentList();
-        cleanResults();
-        update();
-        resetCount();
-    }
-
-    private void resetCount() {
-        LogHelper.d(TAG, "Download Count: " + HentoidApp.getDownloadCount());
-        HentoidApp.setDownloadCount(0);
-        NotificationManager manager = (NotificationManager) mContext.getSystemService(
-                Context.NOTIFICATION_SERVICE);
-        manager.cancel(0);
-    }
-
-    private void initUI(View rootView) {
-        mListView = (RecyclerView) rootView.findViewById(R.id.list);
-        loadingText = (TextView) rootView.findViewById(R.id.loading);
-        emptyText = (TextView) rootView.findViewById(R.id.empty);
-
-        mListView.setHasFixedSize(true);
-        llm = new LinearLayoutManager(mContext);
-        mListView.setLayoutManager(llm);
-
-        mAdapter = new ContentAdapter(mContext, result, this);
-        mListView.setAdapter(mAdapter);
-
-        if (mAdapter.getItemCount() == 0) {
-            mListView.setVisibility(View.GONE);
-            loadingText.setVisibility(View.VISIBLE);
-        }
-
-        mDrawerLayout = (DrawerLayout) getActivity().findViewById(R.id.drawer_layout);
-        mDrawerLayout.addDrawerListener(this);
-
-        btnPage = (Button) rootView.findViewById(R.id.btnPage);
-        toolbar = (Toolbar) rootView.findViewById(R.id.downloads_toolbar);
-        toolTip = (LinearLayout) rootView.findViewById(R.id.tooltip);
-        refreshLayout = (SwipeRefreshLayout) rootView.findViewById(R.id.swipe_container);
-    }
-
-    @Override
-    public void onDownloadEvent(DownloadEvent event) {
-
     }
 
     // Validate permissions
-    protected void checkPermissions() {
+    private void checkPermissions() {
         if (Helper.permissionsCheck(getActivity(), ConstsImport.RQST_STORAGE_PERMISSION, true)) {
             queryPrefs();
             checkResults();
@@ -327,10 +187,6 @@ public abstract class DownloadsFragment extends BaseFragment implements ContentL
             }
             permissionChecked = true;
         }
-    }
-
-    protected void reset() {
-        Helper.reset(HentoidApp.getAppContext(), getActivity());
     }
 
     protected void queryPrefs() {
@@ -399,6 +255,385 @@ public abstract class DownloadsFragment extends BaseFragment implements ContentL
 
     protected abstract void checkResults();
 
+    private void reset() {
+        Helper.reset(HentoidApp.getAppContext(), getActivity());
+    }
+
+    @Override
+    public void onPause() {
+        super.onPause();
+
+        clearSelection();
+    }
+
+    @Override
+    public void onSaveInstanceState(Bundle outState) {
+        super.onSaveInstanceState(outState);
+
+        mListState = llm.onSaveInstanceState();
+        outState.putParcelable(LIST_STATE_KEY, mListState);
+    }
+
+    @Override
+    public void onViewStateRestored(@Nullable Bundle state) {
+        super.onViewStateRestored(state);
+
+        if (state != null) {
+            mListState = state.getParcelable(LIST_STATE_KEY);
+        }
+    }
+
+    @Override
+    public void onViewCreated(View view, Bundle state) {
+        super.onViewCreated(view, state);
+
+        if (mListState != null) {
+            mListState = state.getParcelable(LIST_STATE_KEY);
+        }
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        setRetainInstance(true);
+        setHasOptionsMenu(true);
+
+        mContext = getContext();
+
+        prefs = HentoidApp.getSharedPrefs();
+
+        settingDir = prefs.getString(Consts.SETTINGS_FOLDER, "");
+
+        order = prefs.getInt(
+                ConstsPrefs.PREF_ORDER_CONTENT_LISTS, ConstsPrefs.PREF_ORDER_CONTENT_ALPHABETIC);
+
+        qtyPages = Integer.parseInt(
+                prefs.getString(
+                        ConstsPrefs.PREF_QUANTITY_PER_PAGE_LISTS,
+                        ConstsPrefs.PREF_QUANTITY_PER_PAGE_DEFAULT + ""));
+    }
+
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, final ViewGroup container,
+                             Bundle savedInstanceState) {
+        View rootView = inflater.inflate(R.layout.fragment_downloads, container, false);
+
+        initUI(rootView);
+        attachScrollListener();
+        attachOnClickListeners(rootView);
+
+        return rootView;
+    }
+
+    private void initUI(View rootView) {
+        mListView = (RecyclerView) rootView.findViewById(R.id.list);
+        loadingText = (TextView) rootView.findViewById(R.id.loading);
+        emptyText = (TextView) rootView.findViewById(R.id.empty);
+
+        mListView.setHasFixedSize(true);
+        llm = new LinearLayoutManager(mContext);
+        mListView.setLayoutManager(llm);
+
+        mAdapter = new ContentAdapter(mContext, result, this);
+        mListView.setAdapter(mAdapter);
+
+        if (mAdapter.getItemCount() == 0) {
+            mListView.setVisibility(View.GONE);
+            loadingText.setVisibility(View.VISIBLE);
+        }
+
+        mDrawerLayout = (DrawerLayout) getActivity().findViewById(R.id.drawer_layout);
+        mDrawerLayout.addDrawerListener(new DrawerLayout.SimpleDrawerListener() {
+
+            @Override
+            public void onDrawerStateChanged(int newState) {
+                mDrawerState = newState;
+                getActivity().invalidateOptionsMenu();
+            }
+        });
+
+        btnPage = (Button) rootView.findViewById(R.id.btnPage);
+        toolbar = (Toolbar) rootView.findViewById(R.id.downloads_toolbar);
+        toolTip = (LinearLayout) rootView.findViewById(R.id.tooltip);
+        refreshLayout = (SwipeRefreshLayout) rootView.findViewById(R.id.swipe_container);
+    }
+
+    protected abstract void attachScrollListener();
+
+    private void attachOnClickListeners(View rootView) {
+        attachPrevious(rootView);
+        attachNext(rootView);
+        attachRefresh(rootView);
+
+        toolTip.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                commitRefresh();
+            }
+        });
+
+        refreshLayout.setEnabled(false);
+        refreshLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
+            @Override
+            public void onRefresh() {
+                commitRefresh();
+            }
+        });
+    }
+
+    private void attachPrevious(View rootView) {
+        ImageButton btnPrevious = (ImageButton) rootView.findViewById(R.id.btnPrevious);
+        btnPrevious.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if (currentPage > 1 && isLoaded) {
+                    currentPage--;
+                    update();
+                } else if (qtyPages > 0 && isLoaded) {
+                    Helper.toast(mContext, R.string.not_previous_page);
+                } else {
+                    LogHelper.d(TAG, R.string.not_limit_per_page);
+                }
+            }
+        });
+    }
+
+    private void attachNext(View rootView) {
+        ImageButton btnNext = (ImageButton) rootView.findViewById(R.id.btnNext);
+        btnNext.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if (qtyPages <= 0) {
+                    LogHelper.d(TAG, R.string.not_limit_per_page);
+                } else {
+                    if (!isLastPage && isLoaded) {
+                        currentPage++;
+                        update();
+                    } else if (isLastPage) {
+                        Helper.toast(mContext, R.string.not_next_page);
+                    }
+                }
+            }
+        });
+    }
+
+    protected abstract void attachRefresh(View rootView);
+
+    @Override
+    public boolean onBackPressed() {
+        // If the drawer is open, back will close it
+        if (mDrawerLayout != null && mDrawerLayout.isDrawerOpen(GravityCompat.START)) {
+            mDrawerLayout.closeDrawers();
+            backButtonPressed = 0;
+
+            return false;
+        }
+
+        if (isSelected) {
+            clearSelection();
+            backButtonPressed = 0;
+
+            return false;
+        }
+
+        if (backButtonPressed + 2000 > System.currentTimeMillis()) {
+            return true;
+        } else {
+            backButtonPressed = System.currentTimeMillis();
+            Helper.toast(mContext, R.string.press_back_again);
+
+            if (llm != null) {
+                llm.scrollToPositionWithOffset(0, 0);
+            }
+        }
+
+        if (!query.isEmpty()) {
+            clearQuery(1);
+        }
+
+        return false;
+    }
+
+    protected void clearQuery(int option) {
+        LogHelper.d(TAG, "Clearing query with option: " + option);
+        if (option == 1) {
+            if (searchView != null) {
+                searchView.clearFocus();
+                searchView.setIconified(true);
+            }
+        }
+        setQuery(query = "");
+        update();
+    }
+
+    private void commitRefresh() {
+        toolTip.setVisibility(View.GONE);
+        refreshLayout.setRefreshing(false);
+        refreshLayout.setEnabled(false);
+        newContent = false;
+        mAdapter.updateContentList();
+        cleanResults();
+        update();
+        resetCount();
+    }
+
+    private void resetCount() {
+        LogHelper.d(TAG, "Download Count: " + HentoidApp.getDownloadCount());
+        HentoidApp.setDownloadCount(0);
+        NotificationManager manager = (NotificationManager) mContext.getSystemService(
+                Context.NOTIFICATION_SERVICE);
+        manager.cancel(0);
+    }
+
+    @Subscribe(threadMode = ThreadMode.MAIN)
+    public void onDownloadEvent(DownloadEvent event) {
+        Double percent = event.percent;
+        if (percent >= 0) {
+            LogHelper.d(TAG, "Download Progress: " + percent);
+        } else if (isLoaded) {
+            showReloadToolTip();
+        }
+    }
+
+    @Override
+    public void onCreateOptionsMenu(final Menu menu, MenuInflater inflater) {
+        inflater.inflate(R.menu.menu_content_list, menu);
+
+        // Associate searchable configuration with the SearchView
+        final SearchManager searchManager = (SearchManager)
+                mContext.getSystemService(Context.SEARCH_SERVICE);
+
+        searchMenu = menu.findItem(R.id.action_search);
+        MenuItemCompat.setOnActionExpandListener(searchMenu,
+                new MenuItemCompat.OnActionExpandListener() {
+                    @Override
+                    public boolean onMenuItemActionExpand(MenuItem item) {
+                        toggleSortMenuItem(menu, false);
+
+                        return true;
+                    }
+
+                    @Override
+                    public boolean onMenuItemActionCollapse(MenuItem item) {
+                        toggleSortMenuItem(menu, true);
+
+                        if (!("").equals(query)) {
+                            query = "";
+                            submitSearchQuery(query, 300);
+                        }
+
+                        return true;
+                    }
+                });
+        searchView = (SearchView) MenuItemCompat.getActionView(searchMenu);
+        searchView.setSearchableInfo(searchManager.getSearchableInfo(
+                getActivity().getComponentName()));
+        searchView.setIconifiedByDefault(true);
+        searchView.setQueryHint(getString(R.string.search_hint));
+        searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
+            @Override
+            public boolean onQueryTextSubmit(String s) {
+                submitSearchQuery(s);
+                searchView.clearFocus();
+
+                return true;
+            }
+
+            @Override
+            public boolean onQueryTextChange(String s) {
+                if (shouldHide && (!s.isEmpty())) {
+                    submitSearchQuery(s, 1000);
+                }
+
+                if (shouldHide && orderUpdated) {
+                    clearQuery(0);
+                    orderUpdated = false;
+                }
+
+                if (!shouldHide && (!s.isEmpty())) {
+                    clearQuery(1);
+                }
+
+                return true;
+            }
+        });
+        SharedPreferences.Editor editor = HentoidApp.getSharedPrefs().edit();
+        if (order == 0) {
+            menu.findItem(R.id.action_order_alphabetic).setVisible(false);
+            menu.findItem(R.id.action_order_by_date).setVisible(true);
+        } else {
+            menu.findItem(R.id.action_order_alphabetic).setVisible(true);
+            menu.findItem(R.id.action_order_by_date).setVisible(false);
+        }
+        // Save current sort order
+        editor.putInt(ConstsPrefs.PREF_ORDER_CONTENT_LISTS, order).apply();
+    }
+
+    @Override
+    public void onPrepareOptionsMenu(Menu menu) {
+        boolean drawerOpen = mDrawerLayout.isDrawerOpen(GravityCompat.START);
+        shouldHide = (mDrawerState != DrawerLayout.STATE_DRAGGING &&
+                mDrawerState != DrawerLayout.STATE_SETTLING && !drawerOpen);
+
+        if (!shouldHide) {
+            MenuItemCompat.collapseActionView(searchMenu);
+            menu.findItem(R.id.action_search).setVisible(false);
+
+            toggleSortMenuItem(menu, false);
+        }
+    }
+
+    private void toggleSortMenuItem(Menu menu, boolean show) {
+        if (order == 0) {
+            menu.findItem(R.id.action_order_by_date).setVisible(show);
+        } else {
+            menu.findItem(R.id.action_order_alphabetic).setVisible(show);
+        }
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case R.id.action_order_alphabetic:
+                cleanResults();
+                orderUpdated = true;
+                order = ConstsPrefs.PREF_ORDER_CONTENT_ALPHABETIC;
+                update();
+                getActivity().invalidateOptionsMenu();
+
+                return true;
+            case R.id.action_order_by_date:
+                cleanResults();
+                orderUpdated = true;
+                order = ConstsPrefs.PREF_ORDER_CONTENT_BY_DATE;
+                update();
+                getActivity().invalidateOptionsMenu();
+
+                return true;
+            default:
+
+                return super.onOptionsItemSelected(item);
+        }
+    }
+
+    private void submitSearchQuery(String s) {
+        submitSearchQuery(s, 0);
+    }
+
+    private void submitSearchQuery(final String s, long delay) {
+        query = s;
+        searchHandler.removeCallbacksAndMessages(null);
+        searchHandler.postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                setQuery(s);
+                cleanResults();
+                update();
+            }
+        }, delay);
+    }
+
     protected void checkContent(boolean clear) {
         if (clear) {
             resetCount();
@@ -446,25 +681,7 @@ public abstract class DownloadsFragment extends BaseFragment implements ContentL
         currentPage = 1;
     }
 
-    @Override
-    public void onResume() {
-        super.onResume();
-
-        checkPermissions();
-
-        if (mListState != null) {
-            llm.onRestoreInstanceState(mListState);
-        }
-    }
-
-    @Override
-    public void onPause() {
-        super.onPause();
-
-        clearSelection();
-    }
-
-    protected void clearSelection() {
+    private void clearSelection() {
         if (mAdapter != null) {
             if (mListView.getScrollState() == RecyclerView.SCROLL_STATE_IDLE) {
                 mAdapter.clearSelections();
@@ -588,32 +805,66 @@ public abstract class DownloadsFragment extends BaseFragment implements ContentL
 
     @Override
     public void onItemSelected(int selectedCount) {
+        isSelected = true;
+        showToolbar(false, true);
 
+        if (selectedCount == 1) {
+            mAdapter.notifyDataSetChanged();
+        }
+
+        if (selectedCount >= 2) {
+            selectTrigger = true;
+        }
+
+        if (mActionMode == null) {
+            mActionMode = toolbar.startActionMode(mActionModeCallback);
+        }
+
+        if (mActionMode != null) {
+            mActionMode.invalidate();
+            mActionMode.setTitle(
+                    selectedCount + (selectedCount > 1 ? " items selected" : " item selected"));
+        }
     }
 
     @Override
     public void onItemClear(int itemCount) {
+        if (mActionMode != null) {
+            if (itemCount >= 1) {
+                mActionMode.setTitle(
+                        itemCount + (itemCount > 1 ? " items selected" : " item selected"));
+            } else {
+                selectTrigger = false;
+                mActionMode.invalidate();
+                mActionMode.setTitle("");
+                mAdapter.notifyDataSetChanged();
+            }
+        }
 
+        if (itemCount == 1 && selectTrigger) {
+            selectTrigger = false;
+
+            if (mActionMode != null) {
+                mActionMode.invalidate();
+            }
+        }
+
+        if (itemCount < 1) {
+            clearSelection();
+            showToolbar(true, false);
+
+            if (mActionMode != null) {
+                mActionMode.finish();
+            }
+        }
     }
 
     @Override
-    public void onDrawerSlide(View drawerView, float slideOffset) {
-        // We don't care about this event.
-    }
-
-    @Override
-    public void onDrawerOpened(View drawerView) {
-        // We don't care about this event.
-    }
-
-    @Override
-    public void onDrawerClosed(View drawerView) {
-        // We don't care about this event.
-    }
-
-    @Override
-    public void onDrawerStateChanged(int newState) {
-//        mDrawerState = newState;
-//        getActivity().invalidateOptionsMenu();
+    public void onContentsWiped() {
+        LogHelper.d(TAG, "All items cleared!");
+        displayNoResults();
+        clearSelection();
+        cleanResults();
+        update();
     }
 }
