+++ app/src/main/java/me/devsaki/hentoid/activities/QueueActivity.java
@@ -8,13 +8,14 @@ import android.view.MenuItem;
 import me.devsaki.hentoid.R;
 import me.devsaki.hentoid.abstracts.BaseActivity;
 import me.devsaki.hentoid.abstracts.BaseFragment;
+import me.devsaki.hentoid.abstracts.BaseFragment.BackInterface;
 import me.devsaki.hentoid.fragments.QueueFragment;
 import me.devsaki.hentoid.util.LogHelper;
 
 /**
  * Handles hosting of QueueFragment for a single screen.
  */
-public class QueueActivity extends BaseActivity implements BaseFragment.BackInterface {
+public class QueueActivity extends BaseActivity implements BackInterface {
     private static final String TAG = LogHelper.makeLogTag(QueueActivity.class);
 
     private BaseFragment baseFragment;
